Renamed wire.proto => submission.proto and put it under 'wire' package.
