+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/Envelope.scala
@@ -6,7 +6,7 @@ package com.daml.ledger.participant.state.kvutils
 import java.util.zip.{GZIPInputStream, GZIPOutputStream}
 
 import com.daml.ledger.participant.state.kvutils.{DamlKvutils => Proto}
-import com.daml.ledger.participant.state.kvutils.WireFormat
+import com.daml.ledger.participant.state.kvutils.wire
 import com.google.protobuf.ByteString
 
 import scala.util.Try
@@ -19,13 +19,13 @@ object Envelope {
 
   sealed trait Message extends Product with Serializable
 
-  final case class SubmissionMessage(submission: WireFormat.DamlSubmission) extends Message
+  final case class SubmissionMessage(submission: wire.DamlSubmission) extends Message
 
   final case class LogEntryMessage(logEntry: Proto.DamlLogEntry) extends Message
 
   final case class StateValueMessage(value: Proto.DamlStateValue) extends Message
 
-  final case class SubmissionBatchMessage(value: WireFormat.DamlSubmissionBatch) extends Message
+  final case class SubmissionBatchMessage(value: wire.DamlSubmissionBatch) extends Message
 
   private val DefaultCompression = true
 
@@ -48,10 +48,10 @@ object Envelope {
         .build
     )
 
-  def enclose(sub: WireFormat.DamlSubmission): Raw.Envelope =
+  def enclose(sub: wire.DamlSubmission): Raw.Envelope =
     enclose(sub, compression = DefaultCompression)
 
-  def enclose(sub: WireFormat.DamlSubmission, compression: Boolean): Raw.Envelope =
+  def enclose(sub: wire.DamlSubmission, compression: Boolean): Raw.Envelope =
     enclose(Proto.Envelope.MessageKind.SUBMISSION, sub.toByteString, compression)
 
   def enclose(logEntry: Proto.DamlLogEntry): Raw.Envelope =
@@ -66,7 +66,7 @@ object Envelope {
   def enclose(stateValue: Proto.DamlStateValue, compression: Boolean): Raw.Envelope =
     enclose(Proto.Envelope.MessageKind.STATE_VALUE, stateValue.toByteString, compression)
 
-  def enclose(batch: WireFormat.DamlSubmissionBatch): Raw.Envelope =
+  def enclose(batch: wire.DamlSubmissionBatch): Raw.Envelope =
     enclose(Proto.Envelope.MessageKind.SUBMISSION_BATCH, batch.toByteString, compression = false)
 
   def open(envelopeBytes: Raw.Envelope): Either[String, Message] =
@@ -96,13 +96,13 @@ object Envelope {
           parseMessageSafe(() => Proto.DamlLogEntry.parseFrom(uncompressedMessage))
             .map(LogEntryMessage)
         case Proto.Envelope.MessageKind.SUBMISSION =>
-          parseMessageSafe(() => WireFormat.DamlSubmission.parseFrom(uncompressedMessage))
+          parseMessageSafe(() => wire.DamlSubmission.parseFrom(uncompressedMessage))
             .map(SubmissionMessage)
         case Proto.Envelope.MessageKind.STATE_VALUE =>
           parseMessageSafe(() => Proto.DamlStateValue.parseFrom(uncompressedMessage))
             .map(StateValueMessage)
         case Proto.Envelope.MessageKind.SUBMISSION_BATCH =>
-          parseMessageSafe(() => WireFormat.DamlSubmissionBatch.parseFrom(uncompressedMessage))
+          parseMessageSafe(() => wire.DamlSubmissionBatch.parseFrom(uncompressedMessage))
             .map(SubmissionBatchMessage)
         case Proto.Envelope.MessageKind.UNRECOGNIZED =>
           Left(s"Unrecognized message kind: ${envelope.getKind}")
@@ -115,13 +115,13 @@ object Envelope {
       case msg => Left(s"Expected log entry, got ${msg.getClass}")
     }
 
-  def openSubmission(envelopeBytes: Raw.Envelope): Either[String, WireFormat.DamlSubmission] =
+  def openSubmission(envelopeBytes: Raw.Envelope): Either[String, wire.DamlSubmission] =
     open(envelopeBytes).flatMap {
       case SubmissionMessage(entry) => Right(entry)
       case msg => Left(s"Expected submission, got ${msg.getClass}")
     }
 
-  def openSubmission(envelopeBytes: Array[Byte]): Either[String, WireFormat.DamlSubmission] =
+  def openSubmission(envelopeBytes: Array[Byte]): Either[String, wire.DamlSubmission] =
     open(envelopeBytes).flatMap {
       case SubmissionMessage(entry) => Right(entry)
       case msg => Left(s"Expected submission, got ${msg.getClass}")
