+++ ledger/participant-state/kvutils/src/test/suite/scala/com/daml/ledger/participant/state/kvutils/api/BatchingQueueSpec.scala
@@ -4,7 +4,7 @@
 package com.daml.ledger.participant.state.kvutils.api
 
 import com.daml.ledger.api.testing.utils.AkkaBeforeAndAfterAll
-import com.daml.ledger.participant.state.kvutils.WireFormat.DamlSubmissionBatch.CorrelatedSubmission
+import com.daml.ledger.participant.state.kvutils.wire.DamlSubmissionBatch.CorrelatedSubmission
 import com.daml.ledger.participant.state.v1.SubmissionResult
 import com.google.protobuf.ByteString
 import org.mockito.ArgumentMatchers.any
