+++ ledger/participant-state/kvutils/tools/codec-benchmark/src/benchmark/scala/com/daml/lf/benchmark/BenchmarkWithLedgerExport.scala
@@ -6,8 +6,8 @@ package com.daml.lf.benchmark
 import java.nio.file.{Files, Paths}
 
 import com.daml.bazeltools.BazelRunfiles
-import com.daml.ledger.participant.state.kvutils.WireFormat.DamlSubmission
-import com.daml.ledger.participant.state.kvutils.`export`.ProtobufBasedLedgerDataImporter
+import com.daml.ledger.participant.state.kvutils.wire.DamlSubmission
+import com.daml.ledger.participant.state.kvutils.export.ProtobufBasedLedgerDataImporter
 import com.daml.ledger.participant.state.kvutils.{Envelope, Raw}
 import com.daml.lf.archive.Decode
 import org.openjdk.jmh.annotations.{Param, Scope, Setup, State}
